central/requirements.txt
flask>=2.0.0
kafka-python>=2.0.2
werkzeug
